+++ BakingHome.log.0
@@ -1,0 +1,12 @@
+Oct 07, 2019 2:28:54 PM duke.Launcher init
+INFO: =============================[ Initializing BakingHome ]===========================
+Oct 07, 2019 2:28:54 PM duke.Launcher start
+INFO: Starting BakingHome
+Oct 07, 2019 2:28:54 PM duke.ui.UiManager start
+INFO: Starting UI...
+Oct 07, 2019 9:05:56 PM duke.Launcher init
+INFO: =============================[ Initializing BakingHome ]===========================
+Oct 07, 2019 9:05:56 PM duke.Launcher start
+INFO: Starting BakingHome
+Oct 07, 2019 9:05:56 PM duke.ui.UiManager start
+INFO: Starting UI...
