+++ src/main/java/duke/logic/CommandManager.java
@@ -11,7 +11,7 @@
 //
 //import java.util.ArrayList;
 //import java.util.List;
-//
+
 ///**
 // * Manager class of commands. An undo stack and redo stack is managed by CommandManager.
 // */
