+++ src/main/java/duke/logic/parser/product/AddProductCommand.java
@@ -1,0 +1,5 @@
+package duke.logic.parser.product;
+
+public class AddProductCommand {
+
+}
