+++ src/main/java/duke/logic/parser/product/ProductCommandParser.java
@@ -1,0 +1,7 @@
+package duke.logic.parser.product;
+
+public class ProductCommandParser {
+
+
+
+}
