+++ src/main/java/duke/model/ingredient/IngredientList.java
@@ -1,0 +1,7 @@
+package duke.model.ingredient;
+
+public interface IngredientList {
+
+
+
+}
