+++ src/main/java/duke/storage/recipe/RecipeList.java
@@ -1,14 +1,14 @@
 package duke.storage.recipe;
 
-import duke.entities_decrypted.recipe.Recipe;
+import duke.model.product.Product;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class RecipeList extends ArrayList<Recipe> {
-    private List<Recipe> recipeList = new ArrayList<Recipe>();
+public class RecipeList extends ArrayList<Product> {
+    private List<Product> productList = new ArrayList<Product>();
 
-    public List<Recipe> getRecipeList() {
-        return recipeList;
+    public List<Product> getProductList() {
+        return productList;
     }
 }
