+++ src/main/resources/view/RecipePage.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.*?>
 <?import javafx.scene.text.Font?>
-<AnchorPane prefHeight="591.0" prefWidth="877.0" stylesheets="@../css/recipe.css"
+<AnchorPane prefHeight="591.0" prefWidth="877.0" stylesheets="@../css/product.css"
             xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1">
 
     <children>
@@ -52,7 +52,7 @@
             </AnchorPane>
 
 
-        <AnchorPane fx:id="recipeListPane" prefHeight="566.0" prefWidth="770.0" AnchorPane.bottomAnchor="0.0"
+        <AnchorPane fx:id="productListPane" prefHeight="566.0" prefWidth="770.0" AnchorPane.bottomAnchor="0.0"
                     AnchorPane.leftAnchor="30.0" AnchorPane.rightAnchor="30.0" AnchorPane.topAnchor="0.0"
                     xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1">
                 <children>
@@ -61,8 +61,11 @@
                             <AnchorPane maxHeight="80.0" minHeight="80.0" minWidth="0.0" prefHeight="80.0"
                                         prefWidth="478.0" stylesheets="@../css/duke.css">
                                 <children>
-                                    <Label alignment="CENTER" contentDisplay="CENTER" layoutX="293.0" layoutY="22.0"
-                                           prefHeight="77.0" prefWidth="758.0" text="All Recipes" textAlignment="CENTER"
+                                    <Label fx:id="tableTitle" alignment="CENTER"
+                                           contentDisplay="CENTER"
+                                           layoutX="293.0" layoutY="22.0"
+                                           prefHeight="77.0" prefWidth="758.0" text="All Products"
+                                           textAlignment="CENTER"
                                            AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0"
                                            AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
                                         <font>
@@ -80,7 +83,14 @@
                                             <ScrollPane fitToWidth="true" hbarPolicy="NEVER" prefHeight="379.0"
                                                         prefWidth="815.0" vbarPolicy="NEVER">
                                                 <content>
-                                                    <TableView fx:id="recipeTable" stylesheets="@../css/duke.css">
+                                                    <TableView fx:id="productListTable"
+                                                               stylesheets="@../css/product.css">
+                                                        <columnResizePolicy>
+                                                            <TableView fx:constant="CONSTRAINED_RESIZE_POLICY"/>
+                                                        </columnResizePolicy>
+                                                    </TableView>
+                                                    <TableView fx:id="showProductTable"
+                                                               stylesheets="@../css/product.css">
                                                         <columnResizePolicy>
                                                             <TableView fx:constant="CONSTRAINED_RESIZE_POLICY"/>
                                                         </columnResizePolicy>
